try int64

+++ surf_seis/vibbox.py
@@ -221,7 +221,7 @@ def vibbox_read(fname, param):
         no_channels=H[3]
         # read data
         f.seek(DATA_OFFSET, os.SEEK_SET)
-        A = np.fromfile(f, dtype=np.int32,
+        A = np.fromfile(f, dtype=np.int64,
                         count=BUFFER_SIZE * NUM_OF_BUFFERS)
         A = A.reshape(int(len(A) / no_channels), no_channels)
     # Sanity check on number of channels provided in yaml
@@ -229,7 +229,7 @@ def vibbox_read(fname, param):
         print('Number of channels in config file not equal to number in data')
         return
     A = (2 * VOLTAGE_RANGE * A) - VOLTAGE_RANGE
-    A = A / np.int32(4294967296)
+    A = A / np.int64(4294967296)
     path, fname = os.path.split(fname)
     try:
         # Use derivative of PPS signal to find pulse start
@@ -250,7 +250,7 @@ def vibbox_read(fname, param):
                 median_absolute_deviation(dt))[0][0] + 90000
         print(samp_to_first_full_second)
         print(np.int(1e6 * (1 - (np.float(samp_to_first_full_second) /
-                               FREQUENCY))))
+                                 FREQUENCY))))
         starttime = UTCDateTime(
             np.int(fname[5:9]), np.int(fname[9:11]), np.int(fname[11:13]),
             np.int(fname[13:15]), np.int(fname[15:17]), np.int(fname[17:19]),
